+++ CHANGELOG.md
@@ -15,6 +15,7 @@ for further information about branching and tagging conventions.
 - Passing creatorDetails instead of externalDetails param in the createWithdrawWithAutoConversion
 - Passing creatorDetails instead of details in the assetCreationRequest
 - Passing creatorDetails instead of details in the assetUpdateRequest
+- Passing creatorDetails instead of details in the createSaleCreationRequest
 
 ### Removed
 - Passing keyLevelToSet to CreateChangeRoleRequestBuilder.createChangeRoleRequest